refactor: cleanup mention plugin
