+++ src/web/pmPlugins/mentionPlugin/mentionPluginKey.ts
@@ -1,0 +1,4 @@
+import { PluginKey } from '@tiptap/pm/state';
+import type { TriggerState } from './types';
+
+export const mentionPluginKey = new PluginKey<TriggerState>('mention');
